+++ gdsfactory/components/array_with_via.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from typing import Optional
+from functools import partial
 
 import numpy as np
 
@@ -119,7 +120,7 @@ if __name__ == "__main__":
 
     PDK = get_generic_pdk()
     PDK.activate()
-    via_stack_big = gf.partial(via_stack_factory, size=(30, 20))
+    via_stack_big = partial(via_stack_factory, size=(30, 20))
     # c = array_with_via(columns=3, width=10, via_spacing=20, port_orientation=90)
     c = array_with_via_2d(
         columns=2,